GEODE-3269: Refactoring ShellCommands
