+++ geode-core/src/main/java/org/apache/geode/management/internal/cli/commands/SetVariableCommand.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package org.apache.geode.management.internal.cli.commands;
+
+import org.springframework.shell.core.annotation.CliCommand;
+import org.springframework.shell.core.annotation.CliOption;
+
+import org.apache.geode.management.cli.CliMetaData;
+import org.apache.geode.management.cli.Result;
+import org.apache.geode.management.internal.cli.i18n.CliStrings;
+import org.apache.geode.management.internal.cli.result.ErrorResultData;
+import org.apache.geode.management.internal.cli.result.ResultBuilder;
+
+public class SetVariableCommand implements GfshCommand {
+  @CliCommand(value = {CliStrings.SET_VARIABLE}, help = CliStrings.SET_VARIABLE__HELP)
+  @CliMetaData(shellOnly = true, relatedTopic = {CliStrings.TOPIC_GFSH})
+  public Result setVariable(
+      @CliOption(key = CliStrings.SET_VARIABLE__VAR, mandatory = true,
+          help = CliStrings.SET_VARIABLE__VAR__HELP) String var,
+      @CliOption(key = CliStrings.SET_VARIABLE__VALUE, mandatory = true,
+          help = CliStrings.SET_VARIABLE__VALUE__HELP) String value) {
+    Result result;
+    try {
+      getGfsh().setEnvProperty(var, String.valueOf(value));
+      result =
+          ResultBuilder.createInfoResult("Value for variable " + var + " is now: " + value + ".");
+    } catch (IllegalArgumentException e) {
+      ErrorResultData errorResultData = ResultBuilder.createErrorResultData();
+      errorResultData.addLine(e.getMessage());
+      result = ResultBuilder.buildResult(errorResultData);
+    }
+    return result;
+  }
+}
